specified how to access dynamic route data with asyncData property
Before it was unclear how to reference dynamic route data. The examples used `slug` and `id` which seemed like generic names for different route param data. So my routes appeared to not be updating but it was because I was using `params.slug` when my file was named `_lesson.vue` and I should have been using `params.lesson`.