Add note that async nuxtServerInit must be promise
I had some debugging pain a couple of days ago using `nuxtServerInit` with a callback-style async action.   Eventually I realized that it must return a Promise for anything async so that the nuxt server can wait on it properly.  A quick note in the docs should help others avoid that same mistake.

Feel free to change the wording & formatting to fit your preferred style!