layout/view.vue now ignores the layout attribute
Per https://github.com/nuxt/nuxt.js/issues/1566 it seems layout/view.vue
is now a standalone page that cannot be composed with other layouts due
to conflicts.

I wasn't sure exactly how to handle the translations, so changed the source code
and inserted the comment in English so it doesn't get lost.