Update dokku deployment
When following the guide on dokku deployment I missed the specification how to run the final process that launches the actual application. When this is not set it fails with a message similar to:

```
Cannot find module '/app/dist'
```

This is solved by specifying which command should be run on the dokku deploy.