optim: made loading docs more elaborate
Hello, Nuxt team.
Thanks for the great work with Nuxt docs. Recently I've been trying to setup a custom nuxt loading component with and I found it hard to follow the docs because the flow on this particular was quite discontinuous on how to customize the the loading component. I saw a few issues on stack overflow where some people had this same problem despite the docs containing all the answers they need.

Hopefully this pull request makes the documentation more elaborate and makes more Vue developers smile :D