docs(configuration-build): add yarn before command (#1415)
* docs(configuration-build): add yarn before command

This command doesn't work with npm or npm run it only works with yarn or npx so I think we need to state which commands it works with. As an npm user I would auto do a npm run command and then think it's broken. Feel free to comment if you prefer that the yarn is not added here and if there is a better way of making it more clear

* Update configuration-build.md