Add video to Plugins guide
I created a free tutorial video about Nuxt Plugins, and this commit adds it to the guides.  

The html I used is a modified version of what I found in `/en/guide/installation.md`.  It does not display properly in the preview, but I'm assuming that when it's on the site it'll show up just like the others.  Please correct me if I'm wrong there.

Hopefully readers of the guides will be able to benefit from the video as-is, but if there are changes you'd like me to make, then let me know and I can do a "docs edit" based on your recommendations.